Restructure heapothesys repo
The new organization allows independently maintaine benchmarks to reside
within the heapothesys repo.
