+++ HyperAlloc/src/main/java/com/amazon/corretto/benchmark/hyperalloc/AllocObject.java
@@ -1,4 +1,4 @@
-package com.amazon.corretto.benchmark.heapothesys;
+package com.amazon.corretto.benchmark.hyperalloc;
 
 import java.util.concurrent.ThreadLocalRandom;
