+++ HyperAlloc/src/main/java/com/amazon/corretto/benchmark/hyperalloc/TaskBase.java
@@ -1,4 +1,4 @@
-package com.amazon.corretto.benchmark.heapothesys;
+package com.amazon.corretto.benchmark.hyperalloc;
 
 import java.util.ArrayDeque;
 import java.util.concurrent.Callable;
