+++ HyperAlloc/src/test/java/com/amazon/corretto/benchmark/hyperalloc/HyperAllocTest.java
@@ -1,4 +1,4 @@
-package com.amazon.corretto.benchmark.heapothesys;
+package com.amazon.corretto.benchmark.hyperalloc;
 
 import org.junit.jupiter.api.Test;
 
@@ -6,22 +6,22 @@ import static com.github.stefanbirkner.systemlambda.SystemLambda.catchSystemExit
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
 
-class HypothesysTest {
+class HyperAllocTest {
 
     @Test
     void SimpleRunTest() {
-        Heapothesys.main(new String[]{"-u", "simple", "-d", "5"});
+        HyperAlloc.main(new String[]{"-u", "simple", "-d", "5"});
     }
 
     @Test
     void DefaultRunTypeTest() {
-        Heapothesys.main(new String[]{"-d", "5"});
+        HyperAlloc.main(new String[]{"-d", "5"});
     }
 
     @Test
     void UnknownRunTypeTest() throws Exception {
         int status = catchSystemExit(
-                () -> Heapothesys.main(new String[]{"-u", "unknown", "-a", "5"}));
+                () -> HyperAlloc.main(new String[]{"-u", "unknown", "-a", "5"}));
         assertThat(status, is(1));
     }
 }
